Implement (incomplete) parsing
It should be possible to parse the latest ENSDF version (all datasets).
This does not mean that any valid ENSDF dataset will be parsed
correctly. Especially the evaluation of numbers with uncertanties is
complicated and not as well-defined as one would like. There are
probably some cases where this fails, and there might also be some cases
where it currently does so silently. More (automated) testing is
required.

Comments and References are stored, but not resolved.
So far, I have mostly focused on level and gamma records.
Other record types are (mostly) implemented already but require more
work.