Append uncertainties to corresponding fields
The parser should be able to process this anyway. In some cases,
continuation records overwrite fields from the main record. However,
in that case, the uncertainty is part of the main property field. By
always adding the uncertainty to the main property field, those cases do
not cause any (further) conflicts.